* Fixing warnings. * Using new function math::deg2rad() where appropriate. * Using new function readKeystroke() in Keyboard.h instead of the read(STDIN_FILENO, ...) approach. This makes it more platform-independent. * Fixing compilation issues when compiling on windows.

+++ LapSim Lite/ParticleSystem.h
@@ -24,8 +24,8 @@ struct Particle
     vel_x = vx + spread * (rnd::rand() - 0.5f);
     vel_y = vy + spread * (rnd::rand() - 0.5f);
     g = gravity_acc;
-    pos_x = pos0.c;
-    pos_y = pos0.r;
+    pos_x = static_cast<float>(pos0.c);
+    pos_y = static_cast<float>(pos0.r);
     life_time = life_t;
   }
